fixed errors where "zusatz"-table does not exist Drawer can be closed by tapping menu icon

+++ lib/utils/db.dart
@@ -347,7 +347,9 @@ class LocationsDB {
     String where = "new_or_modified is null";
     await db.delete("daten", where: where);
     await db.delete("images", where: where);
-    await db.delete("zusatz", where: where);
+    if (hasZusatz) {
+      await db.delete("zusatz", where: where);
+    }
   }
 
   static Future<void> deleteZusatz(int nr) async {
